+++ sample.py
@@ -1,8 +1,0 @@
-#!/usr/bin/python3
-
-"""
-以下コマンドで実行。
-./run.sh -f sample.py
-"""
-
-print("hello python")