+++ src/sample-setup.sh
@@ -8,10 +8,16 @@
 #   0.0.1 初期バージョン作成
 ################################################################################
 
-npm run create welcome-to-atcoder
+work_name="welcome-to-atcoder"
+npm run template:python 2 $work_name
 
-cd welcome-to-atcoder
-cat << EOF > welcome-to-atcoder.py
+# 同じ名前の作業ディレクトリがあったらwork_nameが変わる
+# @see create.sh 
+actual_work_name=`ls -1 | grep $work_name | tail -n 1`
+work_name=$actual_work_name
+
+cd $work_name
+cat << EOF > "${work_name}.py"
 # -*- coding: utf-8 -*-
 # 整数の入力
 a = int(input())
@@ -23,24 +29,22 @@ s = input()
 print("{} {}".format(a+b+c, s))
 EOF
 
-cd test
+cd stdin
 cat << EOF > 1.in
 1
 2 3
 test
-EOF
-cat << EOF > 1.out
-6 test
 EOF
 cat << EOF > 2.in
 72
 128 256
 myonmyon
 EOF
+
+cd ../expect
+cat << EOF > 1.out
+6 test
+EOF
 cat << EOF > 2.out
 456 myonmyon
 EOF
-
-rm 3.in 3.out
-
-echo "welcom-to-atcoderをセットアップしました。"